requirements.txt
appdirs==1.4.4
asgiref==3.3.4
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
cloudinary==1.25.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
crc16==0.1.1
distlib==0.3.1
dj-database-url==0.5.0
dj-rest-auth==2.1.5
Django==3.2.3
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
django-location-field==2.1.0
djangorestframework==3.12.4
filelock==3.0.12
Flask==2.0.1
graphviz==0.16
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==20.9
Pillow==8.2.0
pip-review==1.1.0
psycopg2==2.8.6
pycodestyle==2.7.0
pydotplus==2.0.2
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.6
Werkzeug==2.0.1
whitenoise==5.2.0
